scc: Initial import of 3.0.0
https://github.com/boyter/scc

A tool similar to cloc, sloccount and tokei. For counting physical lines of
code, blank lines, comment lines, and physical lines of source code in many
programming languages.

The goal is to be the fastest code counter possible, but also perform COCOMO
calculation like sloccount and to estimate code complexity similar to
cyclomatic complexity calculators. In short one tool to rule them all.